Moved nav logic and adapter to WPMainNavigationView
