+++ WordPress/src/main/res/layout/main_activity.xml
@@ -29,7 +29,7 @@
         android:visibility="gone"
         tools:visibility="visible"/>
 
-    <android.support.design.widget.BottomNavigationView
+    <org.wordpress.android.ui.main.WPMainNavigationView
         android:id="@+id/bottom_navigation"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
